split the Lightingoff of OMVPortTracker
