+++ src/main/java/frc/robot/subsystems/interfaces/AdvancedTrackerInterface.java
@@ -56,15 +56,4 @@ public interface AdvancedTrackerInterface {
      */
     public long getLastAdvancedTargetUpdate();
 
-    /**
-     * gets brightness of the LEDs on the port-tracking OpenMV sensor module through RIO and PWM
-     * @return double (0 - 1) 
-     */
-    public double getLEDLevel();
-
-    /**
-     * sets brightness of the LEDs on the port-tracking OpenMV sensor module through RIO and PWM
-     * set double (0 - 1)
-     */
-    public void setLEDLevel(double LEDLevel);
 }
